Title: 사각형 채우기 3, Time: 243ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240402/사각형 채우기 3/rectangle-fill-3.java
@@ -11,13 +11,13 @@ public class Main {
 
         int n = Integer.parseInt(br.readLine());
 
-        int[] dp = new int[1001];
+        long[] dp = new long[1001];
         dp[0] = 1;
         dp[1] = 2;
 
         for(int i = 2; i <= n; i++) {
             dp[i] = ((dp[i-1] * 2) % mod + (dp[i-2] * 3 % mod)) % mod;
-            for(int j = i-3; j >= 0; j--) {
+            for(int j = i - 3; j >= 0; j--) {
                 dp[i] = (dp[i] + dp[j] * 2) % mod;
             }
         }
